Ammended doctests to have the correct input types

+++ mip_convert/mip_convert/plugins/base/data/processors.py
@@ -951,12 +951,12 @@ def eos_insitu(zt, zs, zh):
 
     * From the NEMO `eos_insitu` docstring
 
-        >>> round(eos_insitu(40., 40., 10000.), 5)
+        >>> round(eos_insitu(np.float64(40.), np.float64(40.), np.float64(10000.)), 5)
         np.float64(1060.93299)
 
     * From Jackett and McDougall (1995) [eos_insitu_1]_
 
-        >>> round(eos_insitu(3., 35.5, 3000.), 3)
+        >>> round(eos_insitu(np.float64(3.), np.float64(35.5), np.float64(3000.)), 3)
         np.float64(1041.833)
 
     Notes
